refactor: dialog 컴포넌트 분리(#200)
