+++ src/features/ticket/index.ts
@@ -1,0 +1,1 @@
+export { LoginConfirmationDialog } from './components/LoginConfirmationDialog';
